requirements.txt
dgl-cu111
torch
numpy
pandas
scipy
sklearn
dgllife
